feat: added eth balance check to exercise and withdraw
